server/requirements.txt
aiofiles==0.4.0
httpcore==0.3.0
httptools==0.0.13
sanic==19.9.0
setuptools==40.8.0
websockets==7.0.0
aiopg==1.0.0
psycopg2-binary==2.8.3
SQLAlchemy==1.3.10
marshmallow==3.2.1
bcrypt==3.1.7
cffi==1.13.0
passlib==1.7.1
pycparser==2.19
six==1.12.0
python-dotenv==0.10.3
# pytest
aiohttp==3.6.0
appdirs==1.4.3
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.3.0
attrs==19.1.0
black==19.3b0
chardet==3.0.4
click==7.0
faker==2.0.2
idna==2.8
importlib-metadata==0.23
more-itertools==7.2.0
multidict==4.5.2
packaging==19.1
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.2
pysnooper==0.2.8
pytest-sanic==1.0.0
pytest==5.1.2
python-dateutil==2.8.0
text-unidecode==1.3
toml==0.10.0
wcwidth==0.1.7
yarl==1.3.0
zipp==0.6.0